+++ Notepad2e.vcxproj
@@ -322,7 +322,8 @@
     <ClCompile Include="src\Dialogs.c" />
     <ClCompile Include="src\Dlapi.c" />
     <ClCompile Include="src\Edit.c" />
-    <ClCompile Include="src\Extension\EditHelper.cpp" />
+    <ClCompile Include="src\Extension\EditHelper.c" />
+    <ClCompile Include="src\Extension\EditHelperEx.cpp" />
     <ClCompile Include="src\Extension\HLSelection.c" />
     <ClCompile Include="src\Extension\InlineProgressBarCtrl.c" />
     <ClCompile Include="src\Extension\StrToHex.c" />
@@ -399,6 +400,7 @@
     <ClInclude Include="src\Dlapi.h" />
     <ClInclude Include="src\Edit.h" />
     <ClInclude Include="src\Extension\EditHelper.h" />
+    <ClInclude Include="src\Extension\EditHelperEx.h" />
     <ClInclude Include="src\Extension\HLSelection.h" />
     <ClInclude Include="src\Extension\InlineProgressBarCtrl.h" />
     <ClInclude Include="src\Extension\SciCall.h" />
