+++ src/Extension/HLSelection.h
@@ -1,3 +1,5 @@
+#pragma once
+
 enum SH_PLACE
 {
   SH_INIT,
@@ -10,9 +12,6 @@ typedef	enum
   HL_SE_REJECT = 1 << 1
 }	HL_SELEDIT_STOP_OPT;
 
-#ifdef __cplusplus
-extern "C" {
-#endif
 void	HLS_init();
 void	HLS_release();
 UINT	HLS_Sci_event_mask(BOOL range_not);
@@ -22,6 +21,3 @@ void	HLS_on_notification(int code, struct SCNotification *scn);
 void	HLS_Update_selection(UINT place);
 void	HLS_Edit_selection_start(const BOOL highlightAll);
 BOOL	HLS_Edit_selection_stop(UINT mode);
-#ifdef __cplusplus
-}//end extern "C"
-#endif
